requirements.txt
# Dependencias principales del proyecto procesador_exif
Pillow>=9.0.0
simplekml>=1.3.0
pandas>=1.4.0
openpyxl>=3.0.0
pyproj>=3.3.0
piexif>=1.1.0
xlsxwriter>=3.0.0
# flake8
# black
# ruff
